pl_parking/requirements.txt
adas-tsf>=2.4.1
asammdf==7.4.2
drawsvg==2.3.0
fiona==1.10.0
geopandas==0.14.4
numpy==1.26.4
opencv-python==4.10.0.82
plotly==5.22.0
plotly-express==0.4.1
psycopg2-binary==2.9.9
pyproj==3.6.1
rrecreader==1.25.4.post0
torch==2.4.1
